hajri-engine/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
pendulum>=3.0.0
asyncpg>=0.29.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
